Add new scraping func and renamed file
It is now possible to scrape the whole plan for a program. You can also
scrape the plan for a specific profile in that program.

Courses are stored as:
{"Termin X" : [
	{"Period 1" : [
		{courses...}
		]
	,
	{"Period 2 : [
		{courses...}]
	]

This structure follows the regular schedule at LiU where Courses are
divided into periods and periods into semesters which gives the nesting.